extract compile commands from vACDM.cpp
